requirements.txt
Flask
flask-cors
boto3
python-dotenv
